Inline helper function for chunking in update_cache.
